+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `pipecat.utils.timeout.wait_for()` which uses `wait_for2` package to
+  avoid `asyncio.wait_for()` issues in Python < 3.12.
+
 - Allow passing custom pipeline sink and source processors to a
   `Pipeline`. Pipeline source and sink processors are used to know and control
   what's coming in and out of a `Pipeline` processor.
@@ -45,6 +48,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `FrameProcessor.get_parent()`.
 
 ### Fixed
+
+- Replaced `asyncio.wait_for()` for Pipecat's `wait_for()`. In Python 3.10,
+  `asyncio.wait_for()` has issues regarding task cancellation (i.e. cancellation
+  is never propagated).
+  See https://bugs.python.org/issue42130
 
 - Fixed an `AudioBufferProcessor` issues that would cause audio overlap when
   setting a max buffer size.